requirements.txt
pymongo>=4.0.0
psutil>=5.9.0
matplotlib>=3.0.0
oprnpyxl>=3.1.5
psycopg2-binary-2.9.10